requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cairocffi==1.7.1
celery==5.3.4
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
croniter==6.0.0
cssselect2==0.8.0
Django==5.0.14
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-ratelimit==4.1.0
django-sslserver==0.22
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
instaloader==4.14.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
MarkupSafe==3.0.2
mcp==1.13.1
MouseInfo==0.1.3
multidict==6.6.3
numpy==2.3.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.1.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
pyphen==0.17.2
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==311
redis==4.6.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
rpds-py==0.27.1
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.35.0
sentry-sdk==2.37.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
trio==0.30.0
trio-websocket==0.12.2
twilio==9.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.11
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
wsproto==1.2.0
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
zopfli==0.2.3.post1
